// generated/public/MessageChannels.ts
// @generated
// This file is automatically generated by Kanel. Do not modify manually.

import { type UsersId } from './Users';

/** Identifier type for public.message_channels */
export type MessageChannelsId = string & { __brand: 'MessageChannelsId' };

/** Represents the table public.message_channels */
export default interface MessageChannels {
  id: MessageChannelsId;

  sender_id: UsersId | null;

  receiver_id: UsersId | null;

  created_at: Date;

  updated_at: Date;
}

/** Represents the initializer for the table public.message_channels */
export interface MessageChannelsInitializer {
  id: MessageChannelsId;

  sender_id?: UsersId | null;

  receiver_id?: UsersId | null;

  /** Default value: CURRENT_TIMESTAMP */
  created_at?: Date;

  /** Default value: CURRENT_TIMESTAMP */
  updated_at?: Date;
}

/** Represents the mutator for the table public.message_channels */
export interface MessageChannelsMutator {
  id?: MessageChannelsId;

  sender_id?: UsersId | null;

  receiver_id?: UsersId | null;

  created_at?: Date;

  updated_at?: Date;
}